moving back to jetson

+++ msgs/motors_msg_pb.jl
@@ -27,7 +27,7 @@ function meta(::Type{MOTORS})
     ProtoBuf.metalock() do
         if !isassigned(__meta_MOTORS)
             __meta_MOTORS[] = target = ProtoMeta(MOTORS)
-            allflds = Pair{Symbol,Union{Type,String}}[:front_left => Float64, :front_right => Float64, :back_right => Float64, :back_left => Float64, :time => Float64]
+            allflds = Pair{Symbol,Union{Type,String}}[:front_left => Int32, :front_right => Int32, :back_right => Int32, :back_left => Int32, :time => Float64]
             meta(target, MOTORS, allflds, ProtoBuf.DEF_REQ, ProtoBuf.DEF_FNUM, ProtoBuf.DEF_VAL, ProtoBuf.DEF_PACK, ProtoBuf.DEF_WTYPES, ProtoBuf.DEF_ONEOFS, ProtoBuf.DEF_ONEOF_NAMES)
         end
         __meta_MOTORS[]
@@ -35,13 +35,13 @@ function meta(::Type{MOTORS})
 end
 function Base.getproperty(obj::MOTORS, name::Symbol)
     if name === :front_left
-        return (obj.__protobuf_jl_internal_values[name])::Float64
+        return (obj.__protobuf_jl_internal_values[name])::Int32
     elseif name === :front_right
-        return (obj.__protobuf_jl_internal_values[name])::Float64
+        return (obj.__protobuf_jl_internal_values[name])::Int32
     elseif name === :back_right
-        return (obj.__protobuf_jl_internal_values[name])::Float64
+        return (obj.__protobuf_jl_internal_values[name])::Int32
     elseif name === :back_left
-        return (obj.__protobuf_jl_internal_values[name])::Float64
+        return (obj.__protobuf_jl_internal_values[name])::Int32
     elseif name === :time
         return (obj.__protobuf_jl_internal_values[name])::Float64
     else
